requirements.txt
Django==1.7b1
Pillow==2.3.1
PyJWT==0.1.9
django-cors-headers==0.12
django-custom-user==0.4
django-model-utils==2.0.3
djangorestframework==2.3.13
djangorestframework-jwt==0.1.4
wsgiref==0.1.2
